test update of other fields
